v1.0.0
Version 1.0.0 of Proxmox Installation Setup on Debian 12

The version 1.0.0 of the Proxmox installation setup on Debian 12 provides a practical solution for implementing the Proxmox Virtual Environment (VE) on top of the stable Debian 12 foundation. This set of automated scripts aims to streamline the installation, configuration, and customization of Proxmox, allowing users to quickly establish a high-performance virtualized environment.